more refinement

+++ config.sh
@@ -1,17 +1,21 @@
 
-# TMUX
-rm ~/.tmux.conf
-ln -s ~/dotfiles/.tmux.conf ~/.tmux.conf
+# tmux
+if [[ -e ~/.tmux.conf ]]; then
+ rm ~/.tmux.conf
+ ln -s ~/dotfiles/.tmux.conf ~/.tmux.conf
+fi
 
 # Xconfig
-rm ~/.Xdefaults
+if [[ -e ~/.Xdefaults ]]; then
+ rm ~/.Xdefaults
+fi
+
 if [[ -n $DISPLAY ]]; then
  ln -s ~/dotfiles/.Xdefaults ~/.Xdefaults
  xrdb -merge .Xdefaults
 fi
 
 # Shell
-touch ~/.bash_aliases
-if [ -f ~/.bash_aliases ]; then
+if [ -e ~/.bash_aliases ]; then
  echo "alias vi='vim'" >> ~/.bash_aliases
 fi
